This commit helps into several issues that are described below:
fixes #3 as the app is now cappable of registering and unregistering the user to the routes. Manual testing was done but automated test will happen ocassionaly.

fixes #4 as the app now have map integration and it displays the routes whic the user is suscribed to

continues the work on #10 prviding several doccumentation for many classes